Fix selection of modules in issue jasp-stats/jasp-test-release#2895 (#72)
This should resolve the issue where model names were not specified correctly, and thus the analysis broke and closes jasp-stats/jasp-test-release#2895
It also fixes the issue that after changing code in the .createModelListHelper function, the .predanFuturePredictionPlot broke to some extent.

This is a regression issue, so perhaps the problem was with the qml elements? I am not sure why it broke.